Add ExecutionResultProblem component
